pyproject.toml
[project]
name = "financial-rag-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "load-dotenv>=0.1.0",
    "openai>=1.101.0",
    "pandas>=2.3.2",
    "pypdf>=6.0.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.0",
    "sqlalchemy>=2.0.43",
    "transformers>=4.55.3",
    "uvicorn>=0.35.0",
]
